voice-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
deepgram-sdk==3.2.7
google-generativeai==0.3.2
aiohttp==3.9.1
pydantic==2.5.3
